changing default salt for offline deployments
Using the salt "vdo.ninja" will make it compatible with the native mobile apps. As well, it will avoid problems with users connecting via different IP addresses maybe, as there may not be a singular domain when using on an offline LAN.

I intend to offer a way to change the salt in the native apps eventually, but for now, this is a good approach.

Please be aware also that using this salt with the official VDO.Ninja WSS handshake server could pose a security risk if you expect to use generic stream and room names, like "test" or "steve".  It's only relatively secure if using it with your own custom handshake server.